fix: refactor useUsers
